refactor(Utils): move date stuff to a dedicated file
