+++ src/components/DateChip.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script>
-import utils from '../utils.js'
+import date_utils from '../utils/date.js'
 
 export default {
   props: {
@@ -35,7 +35,7 @@ export default {
   },
   methods: {
     getDateFormatted(dateString) {
-      return utils.prettyDate(dateString)
+      return date_utils.prettyDate(dateString)
     },
     goToDate() {
       if (this.readonly || !this.date) {
