+++ src/components/HistoryCard.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script>
-import utils from '../utils.js'
+import date_utils from '../utils/date.js'
 
 export default {
   props: {
@@ -30,10 +30,10 @@ export default {
   },
   methods: {
     getDateTimeFormatted(dateTimeString) {
-      return utils.offDateTime(dateTimeString)
+      return date_utils.offDateTime(dateTimeString)
     },
     getRelativeDateTimeFormatted(dateTimeString) {
-      return utils.prettyRelativeDateTime(dateTimeString, 'short')
+      return date_utils.prettyRelativeDateTime(dateTimeString, 'short')
     },
   }
 }
