+++ src/views/PriceAddMultiple.vue
@@ -144,7 +144,7 @@ import { mapStores } from 'pinia'
 import { useAppStore } from '../store'
 import api from '../services/api'
 import constants from '../constants'
-import utils from '../utils.js'
+import date_utils from '../utils/date.js'
 
 export default {
   components: {
@@ -179,7 +179,7 @@ export default {
         location_id: null,
         location_osm_id: null,
         location_osm_type: '',
-        date: utils.currentDate(),
+        date: date_utils.currentDate(),
         currency: null,
       },
       productPriceForm: {},
